+++ header.php
@@ -10,6 +10,7 @@
   <link rel="preconnect" href="https://donate.novaramedia.com" crossorigin />
   <link rel="preconnect" href="https://use.typekit.net" crossorigin />
   <link rel="preconnect" href="https://p.typekit.net" crossorigin />
+  <link rel="preconnect" href="https://www.youtube-nocookie.com" crossorigin />
   <link rel="preload" as="style" href="https://use.typekit.net/aki7elm.css" />
   <?php
     get_template_part( 'partials/header/seo' );