+++ partials/front-page/above-the-fold/featured-post-primary.php
@@ -53,7 +53,7 @@ if ( $show_related && ! empty( $meta['_cmb_related_posts'] ) ) {
 if ( $has_embed ) {
   ?>
 <div class="u-video-embed-container background-black">
-  <?php echo render_youtube_embed_iframe( $meta['_cmb_utube'][0], true ); ?>
+  <?php echo render_youtube_embed_iframe( $meta['_cmb_utube'][0] ); ?>
 </div>
   <?php
 } else {