+++ partials/singles/single-post-video.php
@@ -53,7 +53,7 @@
         }
     ?>
       <div class="u-video-embed-container">
-        <?php echo render_youtube_embed_iframe( $meta['_cmb_utube'][0], true, $autoplay ); ?>
+        <?php echo render_youtube_embed_iframe( $meta['_cmb_utube'][0], $autoplay ); ?>
       </div>
     <?php
       } else {